Major schematic update for v2.3
This was my first real PCB project, and looking back with a couple
months of hindsight I made some mistakes. Some best practices for
driving WS2812B strips weren't followed in my original version, and
I've updated the schematic to reflect what I've learned and what
improvements needed to be made.

* Warning added to README about current version
* Schematic reorganized into tiles
* RTC clock removed
* OLED and I2C headers removed
* JTAG footprint updated
* Pin header output replaced with screw terminal
* Hole size for power injection header increased
* Added whole system fuse and PTC fuse to protect ESP32
* Added reverse polarity protection and re-added ability to disconnect
  USB Vbus from Vdd, so it's possible to program when the strip is
  connected to 12V without frying your computer

There are a couple other improvements that could still be made to the
schematic (removing the solderwire connector for example) but it remains to
be seen if I'm going to go back and re-design the PCB to incorporate
these changes or move onto newer projects.